Refactor and fix the legacy example page
