+++ docs/app/(main)/examples/large-data-set/page.mdx
@@ -1,0 +1,8 @@
+import { DataTableExample } from './_datatable-example'
+import { InlineCode } from '@/components'
+
+# Large Data Set
+
+`<DataTable />` can handle large data.
+
+<DataTableExample />
